pyproject.toml
[project]
name = "ai-finance-management"
version = "0.1.0"
description = "FastAPI + PostgreSQL starter with vertical slice architecture, optimized for AI coding"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "fastapi>=0.120.1",
    "pypdf>=6.0.0",
    "pydantic-settings>=2.11.0",
    "python-multipart>=0.0.20",
    "python-dotenv>=1.2.1",
    "sqlalchemy[asyncio]>=2.0.44",
    "structlog>=25.5.0",
    "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.18.2",
    "pyright>=1.1.407",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.2",
]

[tool.ruff]
target-version = "py312"
line-length = 100
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    ".mypy_cache",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort (import sorting)
    "B",      # flake8-bugbear (catch mutable defaults, etc.)
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade (modernize syntax)
    "ANN",    # flake8-annotations (enforce type hints)
    "S",      # flake8-bandit (security)
    "DTZ",    # flake8-datetimez (timezone-aware datetimes)
    "RUF",    # Ruff-specific rules
    "ARG",    # flake8-unused-arguments
    "PTH",    # flake8-use-pathlib (prefer Path over os.path)
]

ignore = [
    "S311",   # Standard random is fine for non-crypto use
    "E501",   # Line too long (formatter handles this)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "ANN", "ARG001", "D", "E731"]
"test_*.py" = ["S101", "ANN", "ARG001", "D", "E731"]
"**/tests/**/*.py" = ["S101", "ANN", "ARG001", "D", "E731"]  # Also apply to nested test directories
"__init__.py" = ["F401"]  # Allow unused imports in package init
"scripts/**/*.py" = ["T201"]  # Allow print in scripts
"app/core/health.py" = ["B008"]  # FastAPI Depends() in defaults is the recommended pattern

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.12"
strict = true
show_error_codes = true
warn_unused_ignores = true

# Practical strictness (not dogmatic)
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false

# Allow test files to have less strict typing
[[tool.mypy.overrides]]
module = "*.tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
asyncio_mode = "auto"  # Creates new event loop per test (proper isolation)
testpaths = ["app", "tests"]
markers = [
    "integration: marks tests requiring real database (deselect with '-m \"not integration\"')",
]

[tool.pyright]
include = ["app"]
exclude = [
    "**/__pycache__",
    ".venv",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache"
]

# Python version configuration
pythonVersion = "3.12"
pythonPlatform = "Darwin"

# Strict mode - enables all strict checks by default
typeCheckingMode = "strict"

# Test files can have unused functions (fixtures, route handlers for testing)
reportUnusedFunction = "none"
